requirements.txt
streamlit
pdfminer.six
google-genai==1.17.0
scikit-learn
dotenv
fpdf2